requirements.txt
einops==0.4.1
matplotlib==3.5.2
numpy==1.22.3
opencv_python==4.6.0.66
Pillow==9.2.0
scipy==1.8.1
timm==0.5.4
torch==1.12.0
torchvision==0.13.0
tqdm==4.64.0
